+++ libs/modal/src/lib/strategies/confirm/index.ts
@@ -1,0 +1,4 @@
+export * from './confirm.component';
+export * from './confirm.module';
+export * from './types';
+export * from './confirm.strategy';