+++ libs/modal/src/lib/wrappers/nz/index.ts
@@ -1,0 +1,3 @@
+export * from './nz-modal-wrapper/nz-modal-wrapper.module';
+export * from './nz-modal-wrapper/nz-modal-wrapper.component';
+export * from './nz-modal-wrapper-factory.service';