+++ libs/notification/src/lib/types.ts
@@ -4,15 +4,7 @@ import { NotificationRef } from './notification-ref';
 
 export interface NotificationConfig {
   timeOut?: number;
-  position?:
-    | 'topLeft'
-    | 'topCenter'
-    | 'topRight'
-    | 'bottomLeft'
-    | 'bottomCenter'
-    | 'bottomRight'
-    | 'topFullWidth'
-    | 'bottomFullWidth';
+  position?: NotificationPosition;
   easing?: string;
   easeTime?: number;
   disableTimeOut?: boolean;
@@ -27,6 +19,13 @@ export interface NotificationContext {
   $implicit: NotificationRef;
 }
 
+export interface NotificationData extends NotificationConfig {
+  type?: NotificationType;
+  title: string | TemplateRef<NotificationContext>;
+  description?: string | TemplateRef<NotificationContext>;
+  closeable?: boolean;
+}
+
 export enum NotificationType {
   Info = 'info',
   Error = 'error',
@@ -34,9 +33,20 @@ export enum NotificationType {
   Success = 'success',
 }
 
-export interface NotificationData extends NotificationConfig {
-  type?: NotificationType;
-  title: string | TemplateRef<NotificationContext>;
-  description?: string | TemplateRef<NotificationContext>;
-  closeable?: boolean;
+export enum NotificationPosition {
+  TopLeft = 'topLeft',
+  TopCenter = 'topCenter',
+  TopRight = 'topRight',
+  BottomLeft = 'bottomLeft',
+  BottomCenter = 'bottomCenter',
+  BottomRight = 'bottomRight',
+  TopFullWidth = 'topFullWidth',
+  BottomFullWidth = 'bottomFullWidth',
+}
+
+export enum NotificationEasing {
+  Linear = 'linear',
+  EaseIn = 'ease-in',
+  EaseOut = 'ease-out',
+  EaseInOut = 'ease-in-out',
 }