+++ libs/table.datasource.http/src/lib/table-datasource-http.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { TableData, TableDataConfig } from '@spryker/table';
+import { TableData, TableDataConfig, TableDatasource } from '@spryker/table';
 import { JsonHttpUrlEncodingCodec } from '@spryker/utils';
 import { Observable } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
@@ -9,7 +9,8 @@ import { TableDatasourceHttpConfig } from './types';
 
 // Add http datasource type to registry
 @Injectable({ providedIn: 'root' })
-export class TableDatasourceHttpService {
+export class TableDatasourceHttpService
+  implements TableDatasource<TableDatasourceHttpConfig> {
   constructor(private http: HttpClient) {}
 
   resolve(