+++ libs/unsaved-changes.guard.drawer/src/index.ts
@@ -1,0 +1,2 @@
+export * from './lib/unsaved-changes-drawer-guard.service';
+export * from './lib/unsaved-changes-drawer-guard.module';