+++ libs/unsaved-changes.monitor.form/src/index.ts
@@ -1,0 +1,2 @@
+export * from './lib/unsaved-changes-form-monitor.directive';
+export * from './lib/unsaved-changes-form-monitor.module';