+++ libs/utils/src/lib/to-boolean.ts
@@ -1,17 +1,7 @@
 import { coerceBooleanProperty } from '@angular/cdk/coercion';
 
-export function ToBoolean() {
-  return (target: Object, key: string, descriptor?: any) => {
-    if (descriptor) {
-      const originalSet = descriptor.set;
-
-      descriptor.set = function(value: boolean | string) {
-        return originalSet.call(this, coerceBooleanProperty(value));
-      };
-
-      return descriptor;
-    }
-
+export function ToBoolean(): PropertyDecorator {
+  return (target, key) => {
     const internalKey = `__${key.toString()}-value`;
 
     const getter = function(this: any) {