+++ libs/table.datasource.http/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.2-next.1](https://github.com/spryker/ui-components/compare/@spryker/table.datasource.http@0.0.2-next.0...@spryker/table.datasource.http@0.0.2-next.1) (2020-09-21)
+
+
+### Bug Fixes
+
+* **notification:** reopen floating notification on inputs change ([ad46b8f](https://github.com/spryker/ui-components/commit/ad46b8fe450c6e35763f00ec2a65e267ee1c289c))
+
+
+
+
+
 ## 0.0.2-next.0 (2020-08-27)
 
 **Note:** Version bump only for package @spryker/table.datasource.http